moduling
tried to pull things out into different modules but didn't succeed - I think I also messed up the process.env somehow
